requirements/agent.txt
grpcio==1.47.0
grpcio-tools==1.47.0
protobuf==3.20.*
numpy
myosuite>=1.2.3
git+https://github.com/aravindr93/mjrl.git
matplotlib
tabulate
